Trait centaurus_core::light::Light

pub trait Light: Deserialize {
    fn hit(&self, position: &Point3<f64>) -> Option<(Vector3<f64>, Rgb<f64>)>;
}

Define functionality for Light objects

Required methods

fn hit(&self, position: &Point3<f64>) -> Option<(Vector3<f64>, Rgb<f64>)>

Calculate from which direction and with which amount of color the position is hit by the current Light.

Loading content...